requirements.txt
Appium-Python-Client==2.11.1
pytest==7.3.1
selenium==4.11.2
pytest-html==3.2.0 